@@ -1,17 +1,17 @@
 name: rules
 description: Rules is a simple yet powerful and feature-rich validation library for both dart and flutter.
-version: 1.2.0+3
+version: 2.0.0-nullsafety.0
 homepage: https://github.com/ganeshrvel/pub-rules
 
 environment:
-  sdk: '>=2.7.0 <3.0.0'
+  sdk: '>=2.12.0-0 <3.0.0'
 
 dependencies:
-  meta: ^1.1.8
+  meta: ^1.3.0-nullsafety.6
 
 dev_dependencies:
   # Linting
-  lint: ^1.1.1
+  lint: ^1.5.1
 
   # test
-  test: ^1.14.5
+  test: ^1.16.0-nullsafety.16
