@@ -1,14 +1,13 @@
-import 'package:meta/meta.dart';
 import 'package:rules/src/helpers/array.dart';
 import 'package:rules/src/helpers/functs.dart';
 
 // rules model
 class RuleModel {
-  String error;
-  List<String> errorList;
+  String? error;
+  List<String?> errorList;
 
   RuleModel({
-    @required this.errorList,
+    required this.errorList,
   }) {
     if (isNotNullOrEmpty(errorList) && isArrayIndexExists(errorList, 0)) {
       error = errorList[0];
@@ -17,6 +16,6 @@ class RuleModel {
     }
 
     error = null;
-    errorList = errorList ?? [];
+    errorList = errorList;
   }
 }
