@@ -16,7 +16,7 @@ void main() {
     });
 
     test('should throw an error', () {
-      Rule rule1;
+      Rule? rule1;
       final rule2 = Rule('', name: 'name', isRequired: true);
       final groupRule = GroupRule([rule1, rule2], name: 'group name');
 
@@ -32,8 +32,8 @@ void main() {
     });
 
     test('should NOT throw an error', () {
-      Rule rule1;
-      Rule rule2;
+      Rule? rule1;
+      Rule? rule2;
       final groupRule = GroupRule([rule1, rule2], name: 'group name');
 
       expect(rule1?.hasError, equals(null));
@@ -42,7 +42,7 @@ void main() {
     });
 
     test('should NOT throw an error', () {
-      Rule rule1;
+      Rule? rule1;
       final rule2 = Rule('abc', name: 'name', isRequired: true);
       final groupRule = GroupRule([rule1, rule2], name: 'group name');
 
